requirements.txt
Adafruit-Blinka==8.41.2
adafruit-circuitpython-busdevice==5.2.9
adafruit-circuitpython-connectionmanager==3.1.0
adafruit-circuitpython-dht==4.0.4
adafruit-circuitpython-requests==4.0.0
adafruit-circuitpython-typing==1.10.3
Adafruit-PlatformDetect==3.69.0
Adafruit-PureIO==1.1.11
certifi==2024.6.2
charset-normalizer==3.3.2
customtkinter==5.2.2
darkdetect==0.8.0
idna==3.7
lgpio==0.2.2.0
packaging==24.1
pyftdi==0.55.4
pyserial==3.5
pyusb==1.2.1
requests==2.32.3
rpi-lgpio==0.6
rpi-ws281x==5.0.0
RPi.GPIO==0.7.1
sysv-ipc==1.1.0
typing_extensions==4.12.2
urllib3==2.2.1
